+++ apps/avatax/src/modules/taxes/get-active-connection-service.ts
@@ -6,6 +6,7 @@ import { ProviderConnection } from "../provider-connections/provider-connections
 import { BaseError } from "../../error";
 import { createLogger } from "../../logger";
 import { err, fromThrowable, ok } from "neverthrow";
+import { ActiveConnectionServiceErrors } from "./get-active-connection-service-errors";
 
 const avataxProviderFactory = ({
   providerConnection,
@@ -19,35 +20,6 @@ const avataxProviderFactory = ({
     authData,
   });
 };
-
-const ActiveConnectionServiceError = BaseError.subclass("ActiveConnectionServiceError");
-
-export const ActiveConnectionServiceErrors = {
-  /**
-   * TODO: What does it mean?  How it should behave?
-   */
-  MissingChannelSlugError: ActiveConnectionServiceError.subclass("MissingChannelSlugError"),
-
-  MissingMetadataError: ActiveConnectionServiceError.subclass("MissingMetadataError"),
-
-  /**
-   * TODO: What does it mean?  How it should behave?
-   * Should it be handled as BrokenConfigurationError?
-   */
-  ProviderNotAssignedToChannelError: ActiveConnectionServiceError.subclass(
-    "ProviderNotAssignedToChannelError",
-  ),
-
-  /**
-   * Will happen when `order-confirmed` webhook is triggered by creating an order in a channel that doesn't use the tax app
-   */
-  WrongChannelError: ActiveConnectionServiceError.subclass("WrongChannelError"),
-
-  BrokenConfigurationError: ActiveConnectionServiceError.subclass("BrokenConfigurationError"),
-} as const;
-
-export type ActiveConnectionServiceErrorsUnion =
-  (typeof ActiveConnectionServiceErrors)[keyof typeof ActiveConnectionServiceErrors];
 
 export function getActiveConnectionService(
   channelSlug: string | undefined,
